Integrated Search Location functionality with fragment and search auto suggestions feature. Used Navigation, ViewModel, LiveData, DataBinding jetpack libraries

+++ app/src/main/java/com/sundeep1501/weatherforecast/backend/data/MWLocation.kt
@@ -3,6 +3,6 @@ package com.sundeep1501.weatherforecast.backend.data
 data class MWLocation(
     val title: String,
     val location_type: String,
-    val woeid: Long,
+    val woeid: Int,
     val latt_long: String
 )
